changed u16 from int to short as per iss01, the code base has to be tested to ensure this change doesn't affect anything else, purely a casting problem

+++ global/global.h
@@ -44,17 +44,17 @@ typedef enum
 	true
 } bool;
 
-typedef unsigned char u8;
-typedef char		  s8;
+typedef unsigned char 	u8;
+typedef char		  	s8;
 
-typedef unsigned int  u16;
-typedef int 		  s16;
+typedef unsigned short 	u16;
+typedef short 		  	s16;
 
 typedef unsigned short 	us16;
 typedef short			ss16;
 
-typedef unsigned int u32;
-typedef int 		 s32;
+typedef unsigned int 	u32;
+typedef int 		 	s32;
 
 /***** type definitions *****/
 /*
